Add files via upload
First version to read in 3D datasets. Works with position vectors only in this version, but continues to have full 2D capability. So you can view positions in 3D while forces are from 2D directions on the 3D points. I wanted to put this version out because it also includes more fixes to the display and inparticular the window/gui updates, which would occasionally freeze the buttons on the display. This is now fixed, enjoy!